backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dnspython==2.7.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
lxml==6.0.0
motor==3.7.1
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.0
pypdf==5.9.0
python-docx==1.2.0
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0

passlib==1.7.4
bcrypt==3.2.2
python-jose

email-validator

python-multipart
